+++ packages/forma-36-react-components/src/components/Workbench/Workbench.stories.tsx
@@ -15,6 +15,7 @@ storiesOf('(alpha)|Workbench', module)
     <Workbench className={text('className', '')} testId={text('testId')}>
       <Workbench.Header
         title={'Page title'}
+        description="Lorem Ipsum dolor sit amet."
         icon={<Icon icon="ArrowDown" />}
         actions={<Button buttonType="muted">Click</Button>}
       />