+++ packages/forma-36-website/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.55](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.54...forma-36-website@0.2.55) (2019-08-27)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.2.54](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.53...forma-36-website@0.2.54) (2019-08-27)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.2.53](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.52...forma-36-website@0.2.53) (2019-08-27)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.2.52](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.51...forma-36-website@0.2.52) (2019-08-27)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.2.51](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.50...forma-36-website@0.2.51) (2019-08-27)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.2.50](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.49...forma-36-website@0.2.50) (2019-08-26)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.2.49](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.48...forma-36-website@0.2.49) (2019-08-26)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.2.48](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.47...forma-36-website@0.2.48) (2019-08-21)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.2.47](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.46...forma-36-website@0.2.47) (2019-08-21)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
 ## [0.2.46](https://github.com/contentful/forma-36/compare/forma-36-website@0.2.45...forma-36-website@0.2.46) (2019-08-08)